refactor: undo 버튼 분리
